Rename multi_armed_bandit.py to bandit.py
