+++ baybe/surrogates/__init__.py
@@ -1,9 +1,9 @@
 """BayBE surrogates."""
 
+from baybe.surrogates.bandit import BernoulliMultiArmedBanditSurrogate
 from baybe.surrogates.custom import CustomONNXSurrogate, register_custom_architecture
 from baybe.surrogates.gaussian_process.core import GaussianProcessSurrogate
 from baybe.surrogates.linear import BayesianLinearSurrogate
-from baybe.surrogates.multi_armed_bandit import BernoulliMultiArmedBanditSurrogate
 from baybe.surrogates.naive import MeanPredictionSurrogate
 from baybe.surrogates.ngboost import NGBoostSurrogate
 from baybe.surrogates.random_forest import RandomForestSurrogate
