+++ baybe/surrogates/bandit.py
